#33 wip split types directory
