+++ src/components/Alert/Alert.types.ts
@@ -1,5 +1,5 @@
 import type { PropsWithChildren } from 'react';
-import type { StatusValue } from '../../types/common';
+import type { StatusValue } from '../../types';
 
 export type AlertProps = PropsWithChildren & {
   status: StatusValue;
