+++ src/features/currency/components/Converter/Converter.types.ts
@@ -1,4 +1,4 @@
-import type { CodeString } from '../../../../types/currency';
+import type { CodeString } from '../../types';
 
 export type ConverterProps = {
   baseAmount: number;
