+++ src/hooks/useTitle.ts
@@ -1,6 +1,6 @@
 import { useMemo } from 'react';
 import { useMatches, type UIMatch } from 'react-router';
-import type { RouteHandle } from '../types/common';
+import type { RouteHandle } from '../types';
 import { useQueryParams } from './useQueryParams';
 
 export const useTitle = (): string => {
