+++ src/pages/ConvertPage/ConvertPage.tsx
@@ -6,8 +6,11 @@ import { EmptyFragment } from '../../features/currency/components/EmptyFragment'
 import { ErrorsFragment } from '../../features/currency/components/ErrorsFragment';
 import { SymbolChips } from '../../features/currency/components/SymbolChips';
 import { extractCurrencyCodes } from '../../features/currency/utils';
+import type {
+  CodeString,
+  Currency,
+} from '../../features/currency/types';
 import { useCurrencies } from '../../stores/currency/currenciesStore';
-import type { CodeString, Currency } from '../../types/currency';
 
 export type ConvertersData = {
   rates: {
